requirements.txt
selenium
beautifulsoup4
xlsxwriter
pandas
python-dotenv
streamlit
groq
openai
